runOn implementation

+++ spec/spec_tests/crud_spec.rb
@@ -1,11 +1,15 @@
 require 'spec_helper'
 
-define_crud_spec_tests('CRUD spec tests', CRUD_TESTS.sort) do |spec, test|
-  let(:client) { authorized_client }
-  let(:collection) { client['crud_spec_test'] }
+describe do
+  define_crud_spec_tests('CRUD spec tests', CRUD_TESTS.sort) do |spec, req, test|
+    let(:client) { authorized_client }
+    let(:collection) { client['crud_spec_test'] }
 
-  before do
-    collection.delete_many
-    test.setup_test(collection)
+    before do
+      if req.nil? || req.satisfied?
+        collection.delete_many
+        test.setup_test(collection)
+      end
+    end
   end
 end
